# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Returns Services API",
        "description": "API documentation for Returns Service App",
        "contact": {
            "email": "admin@adm.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://localhost/api/",
            "description": "Returns App API"
        }
    ],
    "paths": {
...
        "/polaObats": {
            "get": {
                "tags": [
                    "PolaObat"
                ],
                "summary": "getPolaObatList",
                "description": "Get all PolaObats",
                "operationId": "abff72e7c87b3774c4a93467b4969ce0",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "PolaObat"
                ],
                "summary": "createPolaObat",
                "description": "Create PolaObat",
                "operationId": "dbadcd5249dd7b708a8e83a500a89d00",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    ""
                                ],
                                "properties": {
                                    "name": {
                                        "description": "desc",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/polaObats/{id}": {
            "get": {
                "tags": [
                    "PolaObat"
                ],
                "summary": "getPolaObatItem",
                "description": "Get PolaObat",
                "operationId": "8fa50fa7a77bee27ba50064efa50707b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of PolaObat",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "put": {
                "tags": [
                    "PolaObat"
                ],
                "summary": "updatePolaObat",
                "description": "Update PolaObat",
                "operationId": "3622d4c72468a6edc011948bf0504c23",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of PolaObat",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    ""
                                ],
                                "properties": {
                                    "name": {
                                        "description": "desc",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "PolaObat"
                ],
                "summary": "deletePolaObat",
                "description": "Delete PolaObat",
                "operationId": "b745eaa95dce41c61d468a5983618a42",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of PolaObat",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
...
    },
...
}